requirements.txt
Flask==3.1.0
google-cloud-translate==3.18.0
python-dotenv==1.0.1